requirements.txt
affine==2.4.0
aiobotocore==2.9.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.33.2
boto3-stubs==1.34.5
botocore==1.33.2
botocore-stubs==1.34.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
frozenlist==1.4.1
fsspec==2023.12.2
h5py==3.10.0
idna==3.6
jmespath==1.0.1
multidict==6.0.4
mypy-boto3-cloudformation==1.34.0
mypy-boto3-dynamodb==1.34.0
mypy-boto3-ec2==1.34.4
mypy-boto3-lambda==1.34.0
mypy-boto3-rds==1.34.4
mypy-boto3-s3==1.34.0
mypy-boto3-sqs==1.34.0
numpy==1.26.2
pyparsing==3.1.1
pyproj==3.6.1
pystac==1.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
rasterio==1.3.9
requests==2.31.0
s3fs==2023.12.2
s3transfer==0.8.2
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
types-awscrt==0.20.0
types-s3transfer==0.9.0
typing_extensions==4.9.0
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4
